Bugfixes and parallelization (some for only OpenMP 3.0)

+++ tests/run.sh
@@ -11,41 +11,43 @@ make -C ../util mkcromfs unmkcromfs
 
 ## TEST 1
 
-rm -f tmp.cromfs
-../util/mkcromfs a tmp.cromfs -B'*.hh:128' -b16384 -f65536 -e --threads 4 >/dev/null
-rm -rf b
-../util/unmkcromfs tmp.cromfs b >/dev/null
+if true;then
+	rm -f tmp.cromfs
+	../util/mkcromfs a tmp.cromfs -B'*.hh:128' -b16384 -f65536 -e --threads 4 >/dev/null
+	rm -rf b
+	../util/unmkcromfs tmp.cromfs b >/dev/null
 
-( cd a && tar cf - *) | tar tvvf - | sort > a.listing
-( cd b && tar cf - *) | tar tvvf - | sort > b.listing
+	( cd a && tar cf - *) | tar tvvf - | sort > a.listing
+	( cd b && tar cf - *) | tar tvvf - | sort > b.listing
 
-if diff -u a.listing b.listing; then
-	echo "*** TEST 1: PASS"
-else
-	echo "*** TEST 1: FAIL"
+	if diff -u a.listing b.listing; then
+		echo "*** TEST 1: PASS"
+	else
+		echo "*** TEST 1: FAIL"
+	fi
+	rm -rf a.listing b.listing b tmp.cromfs
 fi
-rm -rf a.listing b.listing b tmp.cromfs
-#exit
-
 
 ## TEST 2 (no sparse files)
 
-rm -f tmp.cromfs
-../util/mkcromfs a tmp.cromfs -b64 -f512 --threads 1 > /dev/null
-rm -rf b
-../util/unmkcromfs tmp.cromfs b -s >/dev/null
+if true; then
+	rm -f tmp.cromfs
+	../util/mkcromfs a tmp.cromfs -b64 -f512 --threads 1  >/dev/null
+	rm -rf b
+	../util/unmkcromfs tmp.cromfs b -s >/dev/null
 
-( cd a && tar cf - *) | tar tvvf - | sort > a.listing
-( cd b && tar cf - *) | tar tvvf - | sort > b.listing
+	( cd a && tar cf - *) | tar tvvf - | sort > a.listing
+	( cd b && tar cf - *) | tar tvvf - | sort > b.listing
 
-rm -rf b tmp.cromfs
+	rm -rf b tmp.cromfs
 
-if diff -u a.listing b.listing; then
-	echo "*** TEST 2: PASS"
-else
-	echo "*** TEST 2: FAIL"
+	if diff -u a.listing b.listing; then
+		echo "*** TEST 2: PASS"
+	else
+		echo "*** TEST 2: FAIL"
+	fi
+	rm -rf a.listing b.listing tmp.cromfs b
 fi
-rm -rf a.listing b.listing tmp.cromfs b
 
 ## TEST 3: Boyer-Moore
